@@ -59,7 +59,7 @@
   },
   "type": "commonjs",
   "engines": {
-    "node": ">=13"
+    "node": ">=16.6"
   },
   "watch": {
     "build": {
@@ -92,7 +92,6 @@
     "colorette": "^2.0.16",
     "dot-properties": "^1.0.1",
     "intl-messageformat": "^9.11.1",
-    "html-to-react": "^1.4.7",
     "nookies": "^2.5.2",
     "resolve-accept-language": "^1.0.43"
   },
