@@ -1,14 +1,9 @@
-import IntlMessageFormat from 'intl-messageformat';
 import { useRouter } from 'next/router';
 import { extname, parse as parsePath, resolve } from 'path';
 
 import { highlight, highlightFilePath, log, normalizeLocale } from '../';
 import { BabelifiedMessages } from './babel-plugin';
-import { KeyValueObject } from './properties';
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const HtmlToReactParser = require('html-to-react').Parser;
-const htmlToReactParser = new HtmlToReactParser();
+import { Messages } from './Messages';
 
 /** This is the regular expression to validate message key segments. */
 export const keySegmentRegExp = /^[a-z\d]{1,50}$/i;
@@ -33,17 +28,17 @@ export const TITLE_KEY_ID = 'title';
  * is available first, and if not try to fallback on the `slug`.
  *
  * @param messages - The object containing localized messages of a local scope.
- * @param values - The values (e.g., placeholders) used to format the message, if any.
+ * @param values - The values of the title's placeholders (e.g., `{name: 'Joe'}`), if any.
  *
  * @returns The message message as a string.
  */
-export function getTitle(messages: Messages, values?: MessageValues): string {
-  const titleMessage = messages.format(TITLE_KEY_ID, values, true);
-  const slugMessage = messages.format(SLUG_KEY_ID, values, true);
+export function getTitle(messages: Messages, values?: PlaceholderValues): string {
+  const titleMessage = messages.get(TITLE_KEY_ID);
+  const slugMessage = messages.get(SLUG_KEY_ID);
 
-  const applicableTitle = titleMessage !== '' ? titleMessage : slugMessage;
+  const applicableTitle = titleMessage !== undefined ? titleMessage : slugMessage;
 
-  if (applicableTitle === '') {
+  if (applicableTitle === undefined) {
     log.warn(
       `unable to use ${highlight('getTitle')} in ${highlightFilePath(
         messages.sourceFilePath
@@ -51,9 +46,10 @@ export function getTitle(messages: Messages, values?: MessageValues): string {
         SLUG_KEY_ID
       )} were not found in messages file ${highlightFilePath(messages.messagesFilePath)}`
     );
+    return '';
   }
 
-  return applicableTitle;
+  return applicableTitle.format(values);
 }
 
 /**
@@ -91,10 +87,47 @@ export function getSourceFilePath(messageFilePath: string, sourceFileExtension: 
 }
 
 /**
- * The values (e.g., placeholders) used to format a message.
+ * The value of a message's placeholder (e.g., `{name}`).
  */
-export type MessageValues = {
-  readonly [key: string]: string | number;
+export type PlaceholderValue = string | number;
+
+/**
+ * The values of a message's placeholders (e.g., `{name: 'Joe'}`).
+ */
+export type PlaceholderValues = {
+  [key: string]: string | number;
+};
+
+/**
+ * The value of a message's JSX element (e.g., `<b></b>`).
+ */
+export type JsxValue = JSX.Element;
+
+/**
+ * The values of a message's JSX elements (e.g., `{b: <b></b>}`).
+ */
+export type JsxValues = {
+  [key: string]: JsxValue;
+};
+
+/**
+ * Any (mixed) message value (placeholders and/or JSX).
+ */
+export type MixedValue = PlaceholderValue | JsxValue;
+
+/**
+ * The (mixed) values of a message (placeholder and/or JSX).
+ */
+export type MixedValues = {
+  [key: string]: MixedValue;
+};
+
+/**
+ * Message values by types.
+ */
+export type MessageValuesByType = {
+  placeholderValues: PlaceholderValues;
+  jsxValues: JsxValues;
 };
 
 /**
@@ -105,162 +138,25 @@ export type MessagesIndex = {
 };
 
 /**
- * Object used to format individual localized messages of a local scope.
+ * Type guard to check if a message value is a JSX element.
+ *
+ * @param values - The value of a message (placeholder and/or JSX).
+ *
+ * @returns True is the value is a JSX element, otherwise false.
  */
-export class Message {
-  /** The parent messages object. */
-  private parent: Messages;
-  /** The message key. */
-  readonly key: string;
-  /** The localized message. */
-  private message: string;
-  /** The IntlMessageFormat objet, if required. */
-  private intlMessageFormat: IntlMessageFormat;
-
-  /**
-   * Create an object used to format localized messages of a local scope.
-   *
-   * @param parent - The parent messages object.
-   * @param key - The key of the message.
-   * @param message - The localized message.
-   */
-  constructor(parent: Messages, key: string, message: string) {
-    this.parent = parent;
-    this.key = key;
-    this.message = message;
-  }
-
-  /**
-   * Format a message identified by a key in a local scope.
-   *
-   * @param key - The local scope key identifying the message.
-   * @param values - The values (e.g., placeholders) used to format the message, if any.
-   *
-   * @returns The formatted message as a string.
-   */
-  public format(values?: MessageValues): string {
-    if (values) {
-      this.intlMessageFormat = this.intlMessageFormat
-        ? this.intlMessageFormat
-        : new IntlMessageFormat(
-            this.message.replace(/</g, "'<'").replace(/>/g, "'>'"), // Escape XML for `IntlMessageFormat`.
-            this.parent.locale
-          );
-
-      return String(this.intlMessageFormat.format(values));
-    }
-
-    return this.message;
-  }
+export function isJsxValue(value: MixedValue): value is JsxValue {
+  return !isPlaceholderValue(value);
 }
 
 /**
- * Object used to format localized messages of a local scope.
+ * Type guard to check if a message value is a placeholder.
+ *
+ * @param values - The value of a message (placeholder and/or JSX).
+ *
+ * @returns True is the value is a placeholder, otherwise false.
  */
-export class Messages {
-  /** Localized messages of a local scope. */
-  private messages: Message[] = [];
-  /** An index to optimize `get` access on messages. */
-  private messagesIndex: MessagesIndex = {};
-  /** The current locale from Next.js. */
-  readonly locale: string;
-  /** The source (the file calling `useMessages`) file path. */
-  readonly sourceFilePath: string;
-  /** The messages file path. */
-  readonly messagesFilePath: string;
-
-  /**
-   * Create an object used to format localized messages of a local scope.
-   *
-   * @param keyValueObject - The "key/value" object coming directly from a `.properties` file.
-   * @param locale - The current locale from Next.js.
-   * @param sourceFilePath - The file path of the source file associated with the messages.
-   * @param messagesFilePath - The file path of the messages.
-   */
-  constructor(
-    keyValueObject: KeyValueObject,
-    locale: string,
-    sourceFilePath: string,
-    messagesFilePath: string
-  ) {
-    if (keyValueObject) {
-      Object.keys(keyValueObject).forEach((key) => {
-        this.messagesIndex[key] =
-          this.messages.push(new Message(this, key, keyValueObject[key])) - 1;
-      });
-    }
-    this.locale = normalizeLocale(locale);
-    this.sourceFilePath = sourceFilePath;
-    this.messagesFilePath = messagesFilePath;
-  }
-
-  /**
-   * Format a message identified by a key in a local scope.
-   *
-   * @param key - The local scope key identifying the message.
-   * @param values - The values (e.g., placeholders) used to format the message, if any.
-   * @param suppressWarning - If set to true, will not display a warning message if the key is missing.
-   *
-   * @returns The formatted message as a string.
-   */
-  public format(key: string, values?: MessageValues, suppressWarning = false): string {
-    if (!this.messages.length) {
-      // No need to log the error since it was caught when calling `useMessage()`.
-      return '';
-    }
-
-    const message = this.messages[this.messagesIndex[key]];
-
-    if (message === undefined) {
-      if (!suppressWarning) {
-        log.warn(
-          `unable to format key with identifier ${highlight(key)} in ${highlightFilePath(
-            this.sourceFilePath
-          )} because it was not found in messages file ${highlightFilePath(this.messagesFilePath)}`
-        );
-      }
-      return '';
-    }
-
-    return message.format(values);
-  }
-
-  /**
-   * Format a message identified by a key in a local scope, into JSX.
-   *
-   * @param key - The local scope key identifying the message.
-   * @param values - The values (e.g., placeholders) used to format the message, if any.
-   *
-   * @returns The formatted message into JSX.
-   */
-  public formatJsx(key: string, values?: MessageValues): JSX.Element {
-    if (!this.messages.length) {
-      // No need to log the error since it was caught when calling `useMessage()`.
-      return <></>;
-    }
-
-    const message = this.messages[this.messagesIndex[key]];
-
-    if (message === undefined) {
-      log.warn(
-        `unable to format key with identifier ${highlight(key)} in ${highlightFilePath(
-          this.sourceFilePath
-        )} because it was not found in messages file ${highlightFilePath(this.messagesFilePath)}`
-      );
-      return <></>;
-    }
-
-    return htmlToReactParser.parse(message.format(values)) as JSX.Element;
-  }
-
-  /**
-   * Get all messages contained in a given local scope.
-   *
-   * @returns All messages contained in a given local scope.
-   */
-  public getAll(): Message[] {
-    return this.messages;
-  }
+export function isPlaceholderValue(value: MixedValue): value is PlaceholderValue {
+  return ['string', 'number'].includes(typeof value);
 }
 
 /**
